+++ s3_cloudtrail.tf
@@ -37,3 +37,12 @@ resource "aws_s3_bucket" "cloudtrail_bucket" {
 }
 EOF
 }
+
+resource "aws_s3_bucket_public_access_block" "cloudtrail_bucket" {
+  bucket                  = "${aws_s3_bucket.cloudtrail_bucket.id}"
+  count                   = "${var.trail_bucketname_create}"
+  block_public_acls       = true
+  block_public_policy     = true
+  ignore_public_acls      = true
+  restrict_public_buckets = true
+}